Scrapped serialisers and returned to branching conditional system
