+++ src/systems/scene/SceneSerialiser.cpp
@@ -1,18 +1,14 @@
 #include "SceneSerialiser.h"
 #include "../../utils/StringHelpers.h"
-#include "serialisers/EntitySerialiser.h"
+#include "../../entities/Geometry.h"
+#include "../../entities/Player.h"
+#include "../../entities/SplashScreen.h"
 #include <iostream>
 #include <algorithm>
 
-// Define static members
-GeometrySerialiser SceneSerialiser::geometrySerialiser;
-PlayerSerialiser SceneSerialiser::playerSerialiser;
-SplashScreenSerialiser SceneSerialiser::splashSerialiser;
-std::vector<EntitySerialiser*> SceneSerialiser::serialisers {
-        &geometrySerialiser,
-        &playerSerialiser,
-        &splashSerialiser,
-};
+// Define macros
+#define SEP '|'
+#define DEFINE_FIELD(name, content) name + std::string(":") + content + SEP
 
 std::string SceneSerialiser::Serialise(const std::vector<Entity*>& entities)
 {
@@ -20,9 +16,30 @@ std::string SceneSerialiser::Serialise(const std::vector<Entity*>& entities)
     std::string fileData;
     for (auto entity : entities)
     {
-        // Try find the appropriate serialiser for the given entity
-        auto serialiser = TryGetSerialiser(entity->GetName());
-        if (serialiser) fileData += serialiser->Serialise(entity) + "\n";
+        // Serialise the general entity information
+        fileData += entity->GetName() + SEP;
+        fileData += DEFINE_FIELD("POSITION", StringHelpers::VectorToString(entity->GetPosition()));
+        fileData += DEFINE_FIELD("ROTATION", std::to_string(entity->GetRotation()));
+        fileData += DEFINE_FIELD("Z-INDEX", std::to_string(entity->GetZIndex()));
+
+        // Try find any specific instructions for the given entity
+        std::string entityName = entity->GetName();
+        if (entityName == "Geometry")
+        {
+            auto geometry = dynamic_cast<Geometry*>(entity);
+            fileData += DEFINE_FIELD("DIMENSIONS", StringHelpers::VectorToString(geometry->GetDimensions()));
+            fileData += DEFINE_FIELD("MODEL_PATH", geometry->GetModelData().GetModelPath());
+            fileData += DEFINE_FIELD("TEXTURE_PATH", geometry->GetModelData().GetTexturePath());
+        }
+        else if (entityName == "Player")
+        {
+            auto player = dynamic_cast<Player*>(entity);
+            fileData += DEFINE_FIELD("MODEL_PATH", player->GetModelData().GetModelPath());
+            fileData += DEFINE_FIELD("TEXTURE_PATH", player->GetModelData().GetTexturePath());
+        }
+
+        // End the serialisation entry
+        fileData += "\n";
     }
     return fileData;
 }
@@ -38,18 +55,27 @@ void SceneSerialiser::Deserialise(const std::vector<std::string>& sceneString, O
         // Strip labels from each item
         for (std::string& arg : args) arg = arg.substr(arg.find(':') + 1, arg.size());
 
+        // Get standard entity fields
+        raylib::Vector3 position = StringHelpers::StringToVector(args[ENTITY_POS]);
+        float rotation = std::stof(args[ENTITY_ROT]);
+        int zIndex = std::stoi(args[ENTITY_Z_IDX]);
+
         // Register entities by entity name
-        auto serialiser = TryGetSerialiser(args[ENTITY_NAME]);
-        if (serialiser) entities.push_back(serialiser->Deserialise(args));
+        if (args[ENTITY_NAME] == "Geometry")
+        {
+            raylib::Vector3 dimensions = StringHelpers::StringToVector(args[CUSTOM_FIELD_1]);
+            std::string modelPath = args[CUSTOM_FIELD_2];
+            std::string texturePath = args[CUSTOM_FIELD_3];
+            entities.push_back(new Geometry(position, rotation, dimensions, ModelData(modelPath, texturePath)));
+        }
+        else if (args[ENTITY_NAME] == "Player")
+        {
+            entities.push_back(new Player(position, rotation));
+        }
+        else if (args[ENTITY_NAME] == "SplashScreen")
+        {
+            entities.push_back(new SplashScreen());
+        }
         else std::cout << "\"" << args[ENTITY_NAME] << "\" has no deserialisation protocols defined" << std::endl;
     }
-}
-
-EntitySerialiser* SceneSerialiser::TryGetSerialiser(const std::string& entityName)
-{
-    for (auto& serialiser : serialisers)
-    {
-        if (serialiser->GetName() == entityName) return serialiser;
-    }
-    return nullptr;
 }
