--- src/systems/scene/serialisers/SplashScreenSerialiser.cpp
@@ -1,12 +1,0 @@
-#include "SplashScreenSerialiser.h"
-#include "../../../entities/SplashScreen.h"
-
-std::string SplashScreenSerialiser::Serialise(Entity *entity)
-{
-    return EntitySerialiser::Serialise(entity);
-}
-
-Entity *SplashScreenSerialiser::Deserialise(const std::vector<std::string>& entityArgs)
-{
-    return new SplashScreen();
-}
